+++ sa/CMakeLists.txt
@@ -206,6 +206,7 @@ set(QUERY_SOURCES
   ${QUERY_DIR}/numeric_opfunc.c
   ${QUERY_DIR}/partition.c
   ${QUERY_DIR}/query_aggregate.cpp
+  ${QUERY_DIR}/query_analytic.cpp
   ${QUERY_DIR}/query_cl.c
   ${QUERY_DIR}/query_dump.c
   ${QUERY_DIR}/query_evaluator.c
@@ -225,6 +226,7 @@ set(QUERY_SOURCES
   )
 set(QUERY_HEADERS
   ${QUERY_DIR}/query_aggregate.hpp
+  ${QUERY_DIR}/query_analytic.hpp
   ${QUERY_DIR}/scan_json_table.hpp
   )
 
@@ -364,11 +366,14 @@ set(SESSION_SOURCES
 
 set(XASL_SOURCES
   ${XASL_DIR}/access_json_table.cpp
+  ${XASL_DIR}/xasl_analytic.cpp
   ${XASL_DIR}/xasl_stream.cpp
   )
 set(XASL_HEADERS
   ${XASL_DIR}/access_json_table.hpp
   ${XASL_DIR}/compile_context.h
+  ${XASL_DIR}/xasl_aggregate.hpp
+  ${XASL_DIR}/xasl_analytic.hpp
   ${XASL_DIR}/xasl_stream.hpp
   )
