+++ src/parser/xasl_generation.h
@@ -38,6 +38,15 @@
 #include "regu_var.h"
 #include "xasl.h"
 
+// forward definitions
+
+// *INDENT-OFF*
+namespace cubxasl
+{
+  struct analytic_list_node;
+} // namespace cubxasl
+// *INDENT-ON*
+
 #define MATCH_ALL       1
 
 #define PT_PRED_ARG_INSTNUM_CONTINUE    0x01
@@ -102,7 +111,9 @@ struct aggregate_info
 typedef struct analytic_info ANALYTIC_INFO;
 struct analytic_info
 {
-  ANALYTIC_TYPE *head_list;
+  // *INDENT-OFF*
+  cubxasl::analytic_list_node *head_list;
+  // *INDENT-ON*
   PT_NODE *sort_lists;
   PT_NODE *select_node;
   PT_NODE *select_list;
